+++ src/Facades/Approval.php
@@ -4,13 +4,10 @@ namespace EightyNine\Approvals\Facades;
 
 use Illuminate\Support\Facades\Facade;
 
-/**
- * @see \EightyNine\Approval\Approval
- */
 class Approval extends Facade
 {
     protected static function getFacadeAccessor()
     {
-        return \EightyNine\Approval\Approval::class;
+        return \EightyNine\Approvals\Approval::class;
     }
 }